fix(Proto): 修复心跳消息处理器的类型转换问题
将MessageOperationType.HeartBeat显式转换为byte类型，避免潜在的类型不匹配问题

+++ GameFrameX.Proto/BuiltIn/ActorHeartBeat_-1.cs
@@ -39,7 +39,7 @@ namespace GameFrameX.Proto.BuiltIn;
 /// 请求心跳
 /// </summary>
 [ProtoContract]
-[MessageTypeHandler((-1 << 16) + 1, MessageOperationType.HeartBeat)]
+[MessageTypeHandler((-1 << 16) + 1, (byte)MessageOperationType.HeartBeat)]
 public sealed class ReqActorHeartBeat : MessageObject, IRequestMessage, IHeartBeatMessage
 {
     /// <summary>
@@ -53,7 +53,7 @@ public sealed class ReqActorHeartBeat : MessageObject, IRequestMessage, IHeartBe
 /// 服务器通知心跳结果，因为有些业务需要对心跳结果做处理所以不做成RPC的方式处理
 /// </summary>
 [ProtoContract]
-[MessageTypeHandler((-1 << 16) + 2, MessageOperationType.HeartBeat)]
+[MessageTypeHandler((-1 << 16) + 2, (byte)MessageOperationType.HeartBeat)]
 public sealed class NotifyActorHeartBeat : MessageObject, INotifyMessage, IHeartBeatMessage
 {
     /// <summary>
